requirements.txt
streamlit==1.32.0
protobuf>=3.20,<5
google-generativeai>=0.4.1
python-dotenv==1.0.1
langchain==0.1.12
langchain-google-genai==0.0.11
PyMuPDF==1.23.26
python-magic==0.4.27